WIP: impl basic shape of cmd and cmd args parsing, wire up more commands (stubs)
At this point the shape of the code suggests several measures of compile-time
metaprogramming are necessary to help with runtime metadata, i.e. to wire up
types and static dispatch correctly.

It's probably a good time to consider `method` vs. `proc` w.r.t. to trade-offs
of dynamic dispatch vs. static dispatch. Of course that may not be an answer,
or at least a complete answer, and some measure of metaprogramming will still
be needed to help corral definitions of actions, commands, dispatchers,
parsers, etc.