feat: Add create account and list accounts commands
Closes: #180.
Closes: #177.
Closes: #184.

Accounts can be created with the `/create` or `/createaccount` command. Existing accounts can be listed with the `/list` or `/listaccounts` command.

Account keyuid is now being generated on the multiaccount.

account.loginTimestamp is now NULL when inserted in the DB, and can be updated using `updateAccount` or `updateAccountTimestamp`.

fix: run migrations only on non-existant DBs
Removes the need to pass in a migration definition file. We do not need to run migrations, unless we are instantiating a DB.