ci: limit to static linking in test workflow in GitHub Actions
Using a full matrix of static/shared linking has been useful in the past for
catching edge cases, but none have been experienced in recent months so limit
to static linking to reduce build and run times in this workflow. The full
matrix can always be activated in a PR branch if it's suspected we're bumping
into an edge case as experienced previously.