refactor: remove unnecessary use of cast in modules within examples
A few uses remain:

* In `examples/waku_chat2.nim`; that code was copied from nim-waku and is only
for reference.
* In `chat/client/tasks.nim`, where the signature of
`proc statusContext*(arg: ContextArg)` is intended to match `type Context` of
`examples/chat/task_runner/tasks.nim` but we know the argument passed to it is
actually of `type StatusArg`.
* In multiple modules within `examples/chat/task_runner/`, where casting plays
a role in de/serializing data sent across instances of `AsyncChannel`, and
resolving the correct type of worker (thread or pool) per `kind` metatdata
tracked as part of worker creation.