feat: generate wallet account
Closes: #214, #215.

Almost all the work for #197 is included in this PR.

Generate a new wallet account using `/add`, `/addaccount`

feat: create settings when creating/importing a login account
When creating or importing a new account (a login account, not a wallet account), settings are now created.

feat: list wallet accounts
Wallet accounts can be listed using `/wallets`, `/listwalletaccounts`, and `/listwallets`. This gets all accounts in the user db that are not chat accounts. It would be much better to store all wallet accounts in the db using the `wallet` flag, however status-go did not do this, reserving that flag for the root wallet account and creating a unique db constraint on that column.