refactor: WakuNode.init -> WakuNode.new in status/waku
`status/waku.nim` is an older module that provided just enough waku v2
integration to allow for a test.

This commit modifies it re: the recent nim-waku refactor efforts so that it and
the test for it works correctly.

The module will soon be overhauled by moving the example client's waku
logic (with some adaptations) into `status/waku.nim`.