feat: add exception handling
Closes: #239

Remove all `except Exception as e` where possible. In some instances, transitive dependencies out of our control were the source of this error. Even in instances of the Nim library, like `parseJson`, for example, force us to `except Exception` to prevent it from bubbling up to higher layers in the abstraction.

At the lowest levels, we are catching all errors that inherit from `CatchableError` and translating them to be useful for layers above. At the highest level, which is our public API layer, meant to be directly consumed by consumer clients, we are catching all catchable exceptions from the layers below and translating them in to a Result type.