fix: in getWalletAccount check `chat = FALSE` instead of `wallet = 0`
We're already checking `chat = FALSE` in `getWalletAccounts` (plural), and we
should do the same in `getWalletAccount` (singular).

This allows `getWalletAccount` to retrieve the address for the default created
account, a.k.a. `Status account`. That address should be usable with e.g. the
`/sendtransaction` command.

Also, in SQL statements (and some comments) switch from 0/1 to FALSE/TRUE for
boolean columns.

Closes #289.